more work on editor,
too much problem moving file around after there are created.
no sure what is the obsession of unreal with public and private folder when the project is open source.
this is only cause for aggravations, source and header file end up in different locations.

I start to move all classes to the public folder,